Added unit test for position  on layer calculation
This commits adds unit test for the position on layer calculation in a pixel

+++ spirit-sensor-gateway/common/TypeDefinition.h
@@ -51,8 +51,8 @@ namespace  DataFlow {
     typedef int16_t Speed;
     typedef uint16_t SystemID;
     typedef uint16_t TrackID;
-    typedef float AngleStart;
-    typedef float AngleEnd;
+    typedef double AngleStart;
+    typedef double AngleEnd;
     typedef int PositionOnLayer;
     typedef int Layer;
 };
